+++ src/Application/Resources/Queries/UpdateResourceDefinitions/UpdateResourceDefinitionsQueryHandler.cs
@@ -51,7 +51,7 @@ namespace ResourcePackerGUI.Application.Resources.Queries.UpdateResourceDefiniti
 
                 for (var i = 0; i < request.Resources.Count; i++)
                 {
-                    percentage = (float)(i + 1) / request.Resources.Count * 100f;
+                    percentage = (i + 1f) / request.Resources.Count * 100f;
 
                     var resource = request.Resources[i];
                     if (resource.NameDefined)
